Seed tables. Validacion

+++ src/DAL/MySQL/scripts/createTables/tableRooms.sql
@@ -7,7 +7,7 @@ CREATE TABLE rooms(
     roomNumber INT NOT NULL,
     offer BOOLEAN NOT NULL,
     price INT NOT NULL,
-    discount FLOAT,
+    discount INT,
     cancellation TEXT NOT NULL,
     `status` VARCHAR(255) NOT NULL,
     amenities TEXT NOT NULL
